+++ src/openpulsar/devices/pulsar/__init__.py
@@ -1,12 +1,12 @@
 """Pulsar mouse capability sheets."""
 
-from .xlite_v3_wired import PulsarXliteV3Wired
+from .xlite_wired import PulsarXliteWired
 from .x2a_wired import PulsarX2AWired
 from .x2_wired import PulsarX2Wired
 from .x2h_wired import PulsarX2HWired
 
 __all__ = [
-    "PulsarXliteV3Wired",
+    "PulsarXliteWired",
     "PulsarX2AWired",
     "PulsarX2Wired",
     "PulsarX2HWired",